scrooge-adaptive: Switch the namespaces used by AdaptTrackingDecoder
Problem

Travis-ci scrooge-adaptive tests are failing due to "class not found". In
AdaptTracingDecoder, it used the object's name which ends in a `$` and an
`AdaptSuffix` (or `AdaptDecoderSuffix`) which starts with a `$` to construct the
namespace with double dollar signs. However, we changed the namespace to double
underscores.

Solution

Drop the last `$` in object class name and change the `AdaptSuffix` and
`AdaptDecoderSuffix` to start with double underscores.

Differential Revision: https://phabricator.twitter.biz/D189166