scrooge-generator: Support TypeAgnostic filters in Java generated Service
Problem

It is not currently possible for Java services which implement a generated `ServiceIface`
to easily have user-generated filters applied.

Solution

Update the generated Java `Service` to have a constructor which takes an additional
parameter of type `Filter.TypeAgnostic` which is then applied per method when generating
the `methodService`.

Result

Users can supply `TypeAgnostic` filters which will run in the request path of a given
RPC call when the `ServiceIface` is served via a wrapping `Service`.

JIRA Issues: CSL-4259

Differential Revision: https://phabricator.twitter.biz/D235709