scrooge: Update slf4j test dependencies
Problem

We specify an slf4j logging implementation (log4j) in sbt
but not pants. Additionally, we recommend logback as a defacto
slf4j implementation so usage of log4j is inconsistent with that
recommendation. Lastly, an actual implementation is not necessary
since it is not material to the functioning of the tests. One just needs
to be present to stop the slf4j warning that it could not locate
a binding for the API.

Solution

Update to use `slf4j-nop`, make sure we consistently specify
also in pants.

JIRA Issues: CSL-7166

Differential Revision: https://phabricator.twitter.biz/D246451