scrooge-generator: Don't use adapted args when calling Service.apply
Problem

Currently the generated code for FinagleService utilizes adapted args when
calling finagle.Service#apply. For many people this generates a warning
due to -Ywarn-adapted-args, and for some (ourselves included) causes the
compilation to fail due to -Yno-adapted-args.

Solution

Explicitly wrap arguments in a 2-tuple. We also set -Yno-adapted-args
in order to avoid this in the future.

Result

This fixes #277

Signed-off-by: Kevin Oliver <koliver@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D254337