scrooge-generator: Remove auxiliary constructor from finagleClient.mustache
Problem:
We want to reduce the amount of scrooge generated code to improve its compile time.

Solution:
This constructor that does not use RichClientParam in finagleClient.mustache is unused, so it does not need to be generated.
I removed it along with the corresponding lines in the gold file tests

Result:
Can no longer use that constructor for finagle clients, though it wasn't in use anyway.

JIRA Issues: CSL-8442

Differential Revision: https://phabricator.twitter.biz/D335831