scrooge-generator: Add first and last keys to values/fields dictionaries in enum/struct
Problem

Mustache doesn't currently provide a lot of flexibility for
manipulating lists. In particular, we want to be able to create an if...else
if...else structure to handle elements in a list differently, but that isn't
possible with the current infrastructure.

Solution

Add first and last attributes to the values/fields dictionaries in
enum/struct respectively. These will be boolean values, which will allow us to
use them the same way we'd use section starters

Result

This shouldn't make any difference in existing code, but it adds
mustache functionality if desired.

For example, for a list of values [a, b, c], you can now write

```
{{#values}}
{{^first}} else {{/first}} if (value == {{value}}) //doSomething
{{/values}}
else //doSomethingElse
```

which should yield

```
if (value == a) //doSomething
else if (value == b) //doSomething
else if (value == c) //doSomething
else //doSomethingElse
```

JIRA Issues: CSL-8479

Differential Revision: https://phabricator.twitter.biz/D340128