scrooge: Add a way to use ThriftStructMetaData without reflection
Problem

`ThriftStructMetaData` uses a lot of reflection and it can be a performance hit
for customers who rely on it.

Solution

Add a way to hydrate the metadata for generated thrift structs, and keep the
original functionality for those who are using `ThriftStructCodec` and
`MetaData` differently than we expected. Also, remove reflection from getValue
in ThriftStructField.

Result

An improvement in performance when using `ThriftStructMetaData` for generated
thrift structs.

JIRA Issues: CSL-8857, CSL-8495

Differential Revision: https://phabricator.twitter.biz/D408703