scrooge: Rename snakeCaseToCamelCase to toDromedaryCase
Problem

There was a string utility in `ThriftStructMetaData` called
`snakeCaseToCamelCase` (previously `toCamelCase`) that was slightly
misleading, letting us believe that it could be easily replaced with
`com.twitter.conversions.StringOps.toCamelCase` which is simply not true.

Solution

Rename `snakeCaseToCamelCase` to `toDromedaryCase` to reduce confusion.

Result

The method `toDromedaryCase` handles scrooge specific scenarios and the scaladoc
reflects this.

```
 - "foo_bar" -> "fooBar"
 - "FOO_BAR" -> "fooBar"
 - "foo_BarBaz" -> "fooBarBaz"
```

JIRA Issues: CSL-9209

Differential Revision: https://phabricator.twitter.biz/D412283