scrooge-linter: persistence rule to support recursive references
Problem:
Before Thrift supported nested structures, the TypeResolver could
guaruntee that after resolving a document, there would be no more
ReferenceTypes. This assumption doesn't hold after the generator was
updated to support recursive references. The Linter was not updated
accordingly. The linter also does not enforce Enums to be persisted if
referenced but a persisted struct. I believe this is a mistake.

Solution:
Update the Persistence LintRule accordingly.

Result:
Removed some dead code in the TypeResolver that was introduced when supporting
recursive reference. The structMap is not needed

support recursive references for the persistence rule

JIRA Issues: CSL-9286

Differential Revision: https://phabricator.twitter.biz/D431714