scrooge-generator: Fixed up TypeResolver to preload a bunch of names
Problem

Sometimes we want to define an identifier later in a file, and use it earlier
in the file, for ease of use.  This is possible in Thrift, but not possible
in Scrooge (except for structs).

Solution

Add support for defining other kinds of identifiers later in a file.

JIRA Issues: CSL-9403

Differential Revision: https://phabricator.twitter.biz/D441871