scrooge-linter: Fix the warning in the too many fields lint rule to exclude unions
Problem

We want to caution people against having too many fields in cases where it would
create a problem for java, especially when those fields are arguments in a
method in the generated code. One of the things we caution on is unions, but
there's no method that takes an argument for each union field, since we don't
expect to have more than one field in a union, unlike for things like structs
or exceptions.

Solution

Stop warning on too many fields for unions.

JIRA Issues: CSL-9389

Differential Revision: https://phabricator.twitter.biz/D442726