scrooge-generator: Add a warning when it's ambiguous which file scrooge should look at
Problem

Some customers were confused by the exception that Scrooge threw when it saw two
files with the same name on the include path.

Solution

Add a warning for when there are two importers that can both find a file that's
being included in a Scrooge IDL file.  In addition, deduplicate importers to
simplify this implementation and make resolving filenames more efficient.

Result

There's now a warning when there are two files on the include path with the
same name.  This might help catch issues before they become a problem.

JIRA Issues: CSL-9525

Differential Revision: https://phabricator.twitter.biz/D453784