dodo: Add back in scrooge-generator 2.10.7 publishing
Problem

We're no longer publishing 2.10.7 scrooge-generator, which is needed for old-timey
sbt.  We did this to avoid double-publishing the 2.11 and 2.12 scrooge-generator,
but we now don't publish 2.10, which was an unintended side effect.

We also remove scrooge-generator from the list of published projects, since we
publish it separately.

Solution

Add back in publishing scrooge-generator.

JIRA Issues: CSL-9848

Differential Revision: https://phabricator.twitter.biz/D496345