scrooge: Split `ThriftStructMetaData` Into Unique Cases
Problem / Solution

ThriftStructMetaData is really two classes in one, one that provides
information about fields and annotations via the concrete information
passed in via constructor arguments, and one that does so via reflection
based on the `ThriftStructCodec`. Let's split the two of them into
subclasses of an abstract `ThriftStructMetaData` instead in order to
have a better separation of concerns between the two use cases, and
allow for an additional extension.

JIRA Issues: CSL-9953

Differential Revision: https://phabricator.twitter.biz/D514638