scrooge-core: Change `ThriftStructMetaData` from Abstract Class to Trait
Problem / Solution

`ThriftStructMetaData` was recently changed from a concrete class to an
abstract class. However, the abstract class contains no implementation
and instead should have been a trait. Changing the abstract class to a
trait allows it to be mixed in to existing classes which already extend
a different class. This is helpful for generated classes that we want to
treat as implementations of `ThriftStructMetaData`.

JIRA Issues: CSL-9485

Differential Revision: https://phabricator.twitter.biz/D529101