scrooge-core: Move `ThriftStructMetaDataUtil` to internal namespace
Problem / Solution

We want to start generating implementations of `ThriftStructMetaData`
for generated types, so that they can be used along with the associated
`StructBuilder` for that `ThriftStruct` to have richer APIs for generic
processing. This will necessitate calling `ThriftStructMetaDataUtil`
from generated code. To allow that, but discourage use in other
scenarios, we make it public, but move it to the
`com.twitter.scrooge.internal` namespace.

JIRA Issues: CSL-9485

Differential Revision: https://phabricator.twitter.biz/D529080