scrooge: Reduce the size of generated Scala code
Problem

The generated Scala code has a lot of repetition, both within classes
and across classes. This is wasteful for both binary size and compile
times.

Solution

A variety of mechanisms are employed to reduce code size. The primary
ones include:

 * Creating `com.twitter.scrooge.internal.TProtocols` as a shared
   class which is used for serializing and deserializing.

 * No longer generating a Decoder object for unions.

Result

Less generated code and smaller class files. A side benefit is that
this part of serde code now gets unit tests instead of relying on
end-to-end integration tests.

This also tees up future work to add more smarts to the serde code. An
example of this in the existing code is the optimization for
serializing IndexedSeq's. Before this commit, adding similar code lead
to bloat. Now, a similar change only increases the size of the one
class in an insignificant way.

Differential Revision: https://phabricator.twitter.biz/D527920