scrooge-generator: Scala struct Proxy no longer created by default
Problem

Scala's generated code for structs includes a generated `Proxy` class which
is not commonly used. Everyone pays the cost of this generated code
even if they are not using it.

Solution

Switch the default behavior to no longer generate. If this code is still
desired, it can be opted into on a struct-by-struct basis by adding a Thrift annotation to a struct,
 `(com.twitter.scrooge.scala.generateStructProxy = "true")`.

Result

Less generated code for the common case.

Differential Revision: https://phabricator.twitter.biz/D535795