scrooge-generator: Fix warning in generated code from the cocoa generator
Problem

The cocoa generator outputs enums as NSIntegers which are 32-bit or 64-bit depending
on the cpu architecture, and thrift enums are 32bit. This causes a warning when calling
encode when compiling for 64-bit.

Solution

Added a cast to when enum fields are encoded in the generated models.

JIRA Issues: CEI-412

Differential Revision: https://phabricator.twitter.biz/D539592