scrooge: Make options parser a separate class
==== Problem ====
When using Scrooge as a library there is no way to reuse option parsing. This leads to duplication of the code and extra effort to keep in sync option parsing when using scrooge as a library.

==== Solution ====
Make option parsing a separate class. Parse all the options into a Config.
The Compiler constructor takes Config as its argument.

==== Result ====
Option parsing can be reused.
A Compiler is created only if Option Parsing was successful.
Config class allows easier customization of Compiler default fields.

JIRA Issues: DPB-12912

Differential Revision: https://phabricator.twitter.biz/D561738