scrooge: Java throws an exception when field types don't match
Problem

The scrooge generated scala code throws an exception when reading a struct that
has different field types from that of the expected struct. The Java
functionality doesn't match this, which creates inconsistencies and confusion.

Solution

The generated java code should also throw an exception when it is decoding a
struct that has fields of different types than the  expected struct.

Result

The generated java and scala code now match in functionality.

JIRA Issues: CSL-10274

Differential Revision: https://phabricator.twitter.biz/D565373