scrooge: Preallocate containers in deepCopy constructor
Problem

deepCopy operation is inefficient as we do not preallocate container sizes in advance.
As a result, we reallocate the underlying storage when new elements get inserted into the container.

Solution

Estimate the correct size for the container and use it in the constructor. Size estimate for the new container is:
* old.size() in case of List
* old.size() / 0.75 + 1 in case of Map/Set; this assumes default load factor.

Result

Generated code uses constructors with size estimates.

JIRA Issues: CSL-10391

Differential Revision: https://phabricator.twitter.biz/D590776