scrooge: Preallocate to 4/3*N rather than 2*N on read
Problem

When deserializing thrift, we do preallocate the containers to 2 *
expected_size. This is an overshoot, as capacity needed to store N
elements in a hash map/ hash set with default load factor is better
estimated with 4/3 * N + 1.
As we overallocate, we spend more CPU cycles on allocation + spend more RAM + have worse RAM locality.

Solution

Preallocate to a better estimate.

Result

Observe the correct code generation in unit tests and cpu/latency wins
in several clients of the library.

JIRA Issues: CSL-10534

Differential Revision: https://phabricator.twitter.biz/D599495