scrooge: Add setting to configure root dir importer
Problem
As described in twitter/scrooge/#341 a lot of warnings are generated
when using scrooge in projects with sbt sub modules as a root dir
importer is always added.

Solution
Add a new setting scroogeThriftIncludeRoot that makes this behaviour
configurable. The default behaviour is as it is now.

Result
Users won't see a ton of warnings.

Closes twitter/scrooge#345

Signed-off-by: Joy Bestourous <jbestourous@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D633782