twitter/oss: Unset TRAVIS env var
Problem

We set the `TRAVIS` env var to convince the various `build.sbt` files
to set appropriate configuration in the forked test processes. However,
this is confusing our code coverage tool as it is picking up the env
var and attempting to verify the project as a TRAVIS CI build, which it
is not and thus coverage uploading fails.

Solution

Unset the `TRAVIS` env var before uploading results such that the auto
detection of the CI provider properly picks up the GitHub Action env
vars and does the right thing.

Differential Revision: https://phabricator.twitter.biz/D636932