scrooge-generator: check typeMap from includeMap
Problem

Some linting rules need to traverse the resolved document to check each
definition. In some cases, it fails with `
com.twitter.scrooge.frontend.TypeNotFoundException` because the resolver cannot
resolve the ReferenceType.

TypeResolver relies on a typeMap to resolve ReferenceTypes when definitions are
not defined in order, when the disordered thrift are in included file,
TypeResolver lost the typeMap and cannot resolve the struct.

Solution

Check typeMap from includeMap

JIRA Issues: CSL-11008

Differential Revision: https://phabricator.twitter.biz/D683065