scrooge-core: "validation.min" and "validation.max" should apply to all numeric types
Problem
`validation.min` and `validation.max` only work when
the field type is `Long`, this is because we interpret
the annotation class with the field type. In the Jakarta
definition, the annotation class for the 2 validators is
required to be `Long`, but the field type can be any
numeric values (including Byte, Double, Int, Long, Short).

Solution
Always set the annotation class of `validation.min` and
`validation.max` to be `Long` regardless of the field type.

JIRA Issues: CSL-11318

Differential Revision: https://phabricator.twitter.biz/D742485