[Bazel] [Thrifts] Mark all successful building thrift targets as bazel compat
Problem
======

Targets known to compile and test successfully with bazel should be marked as bazel compatible.

Solution
======

Verified that tests pass

```
./bazel test <targets>/...
```

Mark the targets compatible with

```
./bazel run tools/scripts/mark_compatible -- <targets>::
```

Marking targets as bazel compatible has an implication that if code changes are not compatible with bazel and pants the submit queue will fail. If this becomes problematic or hard to resolve please reach out to #bazel-support.

Please reach out to #bazel-support with any other bazel related questions.

Differential Revision: https://phabricator.twitter.biz/D746026