scrooge: Modify Linter to include error summary
Problem:
Our scrooge linter isn't as clean and easy to
read as it could be. Specifically, our errors
and warnings are intertwined, and it's not
immediately obvious the source of a lint failure

Solution:
Rather than report errors as we find them,
lets report them all at the end as part of an
error summary. We'll also have it print in red
to the console to make it more clear

JIRA Issues: CSL-11354

Differential Revision: https://phabricator.twitter.biz/D776139