scrooge-serializer: Cache maxReusableBufferSize value in ThriftStructSerializer instances
Problem

Profiling serialization-heavy workflows shows `ThriftStructSerializer`
continuously retrieving the value of `maxReusableBufferSize` flag.

Solution

Cache the value of the `maxReusableBufferSize` on `ThriftStructSerializer`
instance creation, treat it as fixed during the instance lifetime. The value of
this flag typically does not change during the run time of an application, so
this was deemed an acceptable tradeoff.

JIRA Issues: CSL-6421

Differential Revision: https://phabricator.twitter.biz/D783669