scrooge-generator: Validate requests when servers receive them
Problem

For services defined in the IDL, we want to validate all
input parameters for all methods defined in the service.
If the parameter is of type struct, union, or exception type,
we want to leverage the validateInstanceValue API.
After we validated all parameters, we will throw a
ThriftValidationException for all validation violations.

Solution

Modified methodService.mustache template for Scala generator
to throw the exception in MethodPerEndpoint definitions

JIRA Issues: CSL-11198

Differential Revision: https://phabricator.twitter.biz/D768362