scrooge: Fixed SwiftGeneratorSpec with SBT.
Summary:
Problem

SwiftGeneratorSpec was failing when being run through SBT due to an NPE
while setting up the tests. Upon further investigation, this NPE was
occuring because the swift test resources were not being included as a
resource directory under SBT, resulting in a null return value from
ClassLoader#getResource().

Solution

These test resource files don't actually need to be under the thrift/
directory, as they are only used in the context of SwiftGeneratorSpec,
so this patch moves them to the standard resource/ directory. This makes
the resources available in tests without having to do any special
resource declaration in build files.