scrooge: Add API to access a codec with a class tag or manifest
Problem

We have a great new API that provides a way to get a codec for a class
(`ThriftStructCodec.forStructClass`), but we recently noticed that many people
use manifests and class tags to get the classes to pass to this API. This is a
great opportunity to expand our API to provide a codec given a
ClassTag/manifest.

Solution

Add an API that returns a codec given a class tag or manifest.

JIRA Issues: CSL-11726

Differential Revision: https://phabricator.twitter.biz/D834052