finagle-ts: scrooge generated models
Problem / Solution

We want scrooge to generate our type definitions and schema models.
This patch implements a plain scrooge `Generator` (not based on
handlebars) and generates structs, enums, and typedefs.

Result

We have a foundation to implement more generated code. We still
need to implement unions, services, constants, and annotations.

Differential Revision: https://phabricator.twitter.biz/D858499