scrooge-generator: ServerValidationMixin should extends MethodPerEndpoint
Problem
The new trait `ServerValidationMixin` is used to provide an alternative
API (prefixed with `violationReturning`) for each method defined in
the IDL. Currently, `ServerValidationMixin` is an independent trait
and we changed the APIs to create a Finagle server to accept
`MethodPerEndpoint extends ServerValidationMixin` as long as
any request args defined in the IDL methods is annotated with `validation.`
annotations. In this case, users have to extends `ServerValidationMixin`
even when they don't want to override the alternative `violationReturning`
APIs.

Solution
Make `ServerValidationMixin` extend `MethodPerEndpoint`.

JIRA Issues: CSL-11764, CSL-11830

Differential Revision: https://phabricator.twitter.biz/D856057