scrooge-core: Define union field info getter for union class
Problem

Some code in Schemer and Scalacheck is using reflection to
retrieve the fieldInfo field from union member classes.

Solution

Add a supported API, `ThriftUnion.fieldInfoForUnionClass`
that works the same as `unionStructFieldInfo` method on an
instance, but it takes the class as a parameter and thus
doesn't require instantiation of the class.

JIRA Issues: CSL-11926

Differential Revision: https://phabricator.twitter.biz/D871986