scrooge-generator: Add @generated annotation to Swift generated code
Problem

In Phabricator, generated code that is checked in can be collapsed by
default in the PR if a `@generated` string is present anywhere in the
text file. Generated code is not subject to code review in the same way
as other code, so we want to improve the UX of code review for our
Swift developers.

Solution

Update the comment header in the swiftgen template to reflect the
new copy, using the `@generated` annotation.

This change is only being made in the swiftgen templates for the time
being, but if others request it, it won't take much to bring it to the
other generators.

Result

The comment header of Swift code generated with Scrooge will have
changed, and Phabricator will recognized the files as generated without
requiring a Phabricator site configuration change.

JIRA Issues: MFC-1378

Differential Revision: https://phabricator.twitter.biz/D879262