Scrooge: support validations on nested fields that are struct, union, and exception
Problem

The nested validations was not working as expected. The expected behavior
when passing `invalidNestedRequest` to `methodPerEndPointClient.validateNestedRequest`
is to throw `TApplicationException` but we get nothing instead. This was because we look for
validation annotations on nestedStructs themselves instead of inspecting deeper in the struct.

Solution

Run validation checks recursively so that If nested fields are passed and their fields are defined with validations annotations, we validate the nestedRequest and return the expected violations.

JIRA Issues: CSL-12029

Differential Revision: https://phabricator.twitter.biz/D911262