scrooge-generator: Removed extra resources/swiftgen BUILD file.
Problem

The BUILD file under resources/swiftgen was causing bazel to not
properly package the necessary swiftgen resources with scrooge, causing
the swift generation to fail at runtime with a
`com.github.mustachejava.MustacheNotFoundException`. In addition, this
BUILD file is extraneous as there is a BUILD file in the parent dir
that includes all necessary resources associated with the scrooge
generators.

Solution

Removed the extraneous BUILD file in swiftgen and just rely on the
parent resources BUILD file to include the swiftgen resources.

Result

Running the swift generator under bazel now succeeds.

JIRA Issues: CEI-1769, DPB-16118

Differential Revision: https://phabricator.twitter.biz/D914924