scrooge: Introduce AnnotatedFieldType to support annotating inner container types
# Problem

Currently, thrift annotations inside of collection types pass the parser, but
are subsequently dropped.

# Solution

This change adds a new `FieldType` – `AnnotatedFieldType`. This is currently
used to support thrift annotations on types inside of containers
(maps/lists/sets). But this can be further extended to apply annotations on any
`FieldType`.

# Result

By doing this, annotations inside of collection types are now propagated and are
available to downstream compilers.

JIRA Issues: CSL-11944

Differential Revision: https://phabricator.twitter.biz/D911997