scrooge-generator-tests: oss sbt compatiblity fix
Problem

we saw failures in sbt tests that  are due to changes we added in
scrooge-generator-tests to make it bazel compatible. This was caused by
how we look up resource files by using systempath which worked for both pants
and bazel but doesn't work for sbt. Since sbt uses classpath to look up resource file.

Solution

Create a condition for bazel to unpack and extract resource from a JAR file
and get the file from the classpath.

JIRA Issues: CSL-12176

Differential Revision: https://phabricator.twitter.biz/D927564