scrooge: compile goldfiles and fix bugs
Problem

We rely on goldfiles to test many features, but we never compile the generated goldfiles.
In generated code, the default value type is ambiguous, but the constructor
requires them to be concrete, it won't compile.

Solution

Fix the bug in goldfiles.
Make it a target.
Have a script to auto-overwrite java and scala goldfiles.

Differential Revision: https://phabricator.twitter.biz/D943474