scrooge-generator: eq check first before iterating over fields in scrooge generated scala code
Problem

For large thrift structs equality checking can result in iterating over
all the fields in the object. For nested structs, this can be quite a lot
of fields. While collections will get a quick `eq` check from their
scala implementation before brute forcing and examining every
element, scrooge generated Structs do not do a referential equality
check first.

Solution

Do a referential equality check `eq` on the structs before iterating
over their fields. This reduces equality check time when structs are
referentially equal, which may not be often, but the check is cheap
and can prevent a lot more work on large structs.

Result

Equality checking on referentially equal scrooge generated scala
types is reduced to `O(1)` from `O(n)` (where n is the number of
fields in the thrift struct and any fields in structs the parent contains,
etc).  This also speeds up checking when a Struct in a Struct is
referentially equal even if the whole thing is not, since it prevents
iterating over the sub-Struct's fields in that case.

Differential Revision: https://phabricator.twitter.biz/D950010