scrooge-generator: provide copyWith methods on structs
Problem
It can be useful to have a setter method for a field instead of having
to call `copy()` (which cannot be type checked across different
structs at compile-time).

Solution
Provide an API to set individual fields on a Thrift struct based on
its name. It's effectively the same thing as `copy()` but only works
on single fields.

Differential Revision: https://phabricator.twitter.biz/D1040974